Standalone mono binding gen (#56)
Large changeset:
- stand-alone c# binding generator
- mono glue registration moved to dynamically loaded plugins.
- use newer c# build tools

A bunch of smaller changes:
* Mark MessageQueue::statistics as const.
* Remove old scons build system files
* Rename export files to be unique between different platforms.
* Use lambda based deferred calls in a few places.
* Normalized windows.h header names
* Support array-like and struct-like properties
* Fixes for using entt reflection + more specific usage types
* Update entt to 3.4.0
* Use arg names from doc data if available.
* Codebase was updated to upstream's 3.2.3 release
* Docs, warnings, build fixes.
* Install built csharp assemblies + remove some logging messages.

+++ scene/resources/world_2d.cpp
@@ -230,7 +230,7 @@ struct SpatialIndexer2D {
             Vector<VisibilityNotifier2D *> added;
             Vector<VisibilityNotifier2D *> removed;
 
-            int visible_cells = (end.x - begin.x) * (end.y - begin.y);
+            uint64_t visible_cells = (uint64_t)(end.x - begin.x) * (uint64_t)(end.y - begin.y);
 
             if (visible_cells > 10000) {
